Helper alternative
I created an interface IRequestAnalyzer so that the code for generating
the info string could be parceled out into different classes
implementing the interface (per OS / CLR Version / Browser etc).

This commit includes the interface, implementations (at least, as of
before Michael's updates)  and an app which is intended to replace
helper itself. It's constructor will take an IEnumerable of the
IRequestAnalyzers so that new implementations can be 'plugged' in on an
as-needed basis. There's a unit test to show how this would work.

As for changes to existing code. I tweaked the names of the constants to
make a little more sense on reading. Hope this works!