Request Analyzer compatability
So what I did here is two pieces,

one, rather than strip out the processing of HttpRequestBase, I added a
method signature to the interface for handling userAgent and browser
capabilities to match Michael's work, but for the implementation I
simply copied my 'mock' classes from the unit test into the lib (under
different names) so that the method overload for Michael's simply
wrapped up those properties and submitted them back into the analyzers
as httprequestss. This maintains compatability and function for both.

The RequestAnalzyerApp now has a method overload for userAgent and
version as well and does the same thing.

I also tweaked the interface for message handling. I now pass the