Better handling of unknown release keys in registry
This commit has the .Net Application checker return the releaseKey from
the registry if it's not a known key tied to a .Net version. The website
then uses this information to tell the user that it can't find the exact
version and suggests they log an issue with the given release key.
This is in response to issue #41: Handle unknown releases better.